Fix slider sample size to 2

+++ services/slider.c
@@ -10,7 +10,7 @@ struct srv_state {
     SENSOR_COMMON;
     uint8_t inited;
     uint8_t pinH, pinL, pinM;
-    uint32_t sample;
+    uint16_t sample;
     uint32_t nextSample;
 };
